// src/app/models/cliente/datos-personales.ts
export class PersonalesModel {
    nombre: string;
    app: string;
    apm: string;
    correo: string;
    fechaNacimiento: Date;
    rfc: string;
    calle: string;
    numero: string;
    colonia: string;
    cp: string;
    estado: string;
    ciudad: string;
    municipio: string;
    terminos: boolean;
}